Former-commit-id: 3de9c639e07894e0c04d3462fc1c5c83216d38ea
Former-commit-id: 6818e523889301a89ccbd7a7f4d55a74ea5d8d7c

+++ src/wid_intro.c
@@ -128,7 +128,7 @@ void wid_intro_visible (void)
 
 static boolean wid_intro_key_event (widp w, const SDL_KEYSYM *key)
 {
-    switch (key->sym) {
+    switch ((int)key->sym) {
         case ' ':
         case SDLK_RETURN:
             wid_intro_play_selected();
